requirements.txt
streamlit
google-generativeai
sentence-transformers
faiss-cpu
wikipedia
pandas
numpy